+++ src/xd/lib/util/bw_test.go
@@ -4,7 +4,6 @@ import "testing"
 
 func TestFormatRate(t *testing.T) {
 
-	rate := float32(1000000.5)
-
+	rate := float64(1000000.5)
 	t.Logf("rate %f %s", rate, FormatRate(rate))
 }
